+++ .codechecker.skiplist
@@ -20,4 +20,4 @@
 -*.pb.cc
 
 # Internal ignore
--*/device/tt_kmd_lib/tt_kmd_lib.c
+-*/umd-uapi/src/umd_uapi.c
