+++ device/CMakeLists.txt
@@ -328,7 +328,7 @@ target_link_libraries(
 
 target_link_libraries(tt-umd PRIVATE umd_asio_lib)
 
-target_link_libraries(tt-umd PRIVATE tt-kmd-lib::tt-kmd-lib)
+target_link_libraries(tt-umd PRIVATE umd-uapi::umd-uapi)
 
 check_function_exists(
     backtrace
