+++ umd-uapi/CMakeLists.txt
@@ -4,20 +4,20 @@
 
 cmake_minimum_required(VERSION 3.25)
 
-# Allow tt-kmd-lib to be built on its own, not only via add_subdirectory().
+# Allow umd-uapi to be built on its own, not only via add_subdirectory().
 if(CMAKE_SOURCE_DIR STREQUAL CMAKE_CURRENT_SOURCE_DIR)
     cmake_minimum_required(VERSION 3.25)
-    project(tt-kmd-lib DESCRIPTION "Tenstorrent kernel-mode driver userspace library" LANGUAGES C)
+    project(umd-uapi DESCRIPTION "Tenstorrent kernel-mode driver userspace library" LANGUAGES C)
     include(GNUInstallDirs)
 endif()
 
-add_library(tt-kmd-lib STATIC)
-add_library(tt-kmd-lib::tt-kmd-lib ALIAS tt-kmd-lib)
+add_library(umd-uapi STATIC)
+add_library(umd-uapi::umd-uapi ALIAS umd-uapi)
 
-target_sources(tt-kmd-lib PRIVATE src/tt_kmd_lib.c)
+target_sources(umd-uapi PRIVATE src/umd_uapi.c)
 
 target_include_directories(
-    tt-kmd-lib
+    umd-uapi
     PUBLIC
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/api>
         $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
@@ -25,9 +25,9 @@ target_include_directories(
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src>
 )
 
-target_compile_features(tt-kmd-lib PRIVATE c_std_11)
+target_compile_features(umd-uapi PRIVATE c_std_11)
 set_target_properties(
-    tt-kmd-lib
+    umd-uapi
     PROPERTIES
         POSITION_INDEPENDENT_CODE
             ON
@@ -35,11 +35,11 @@ set_target_properties(
 
 if(CMAKE_VERSION VERSION_GREATER_EQUAL 3.23)
     target_sources(
-        tt-kmd-lib
+        umd-uapi
         PUBLIC
             FILE_SET api
             TYPE HEADERS
             BASE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/api
-            FILES api/tt-kmd-lib/tt_kmd_lib.h api/tt-kmd-lib/pci_ids.h
+            FILES api/umd-uapi/umd_uapi.h api/umd-uapi/pci_ids.h
     )
 endif()
