+++ umd-uapi/api/umd-uapi/umd_uapi.h
@@ -2,8 +2,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef TTKMD_H_
-#define TTKMD_H_
+#ifndef UMD_UAPI_H_
+#define UMD_UAPI_H_
 
 #include <stddef.h>
 #include <stdint.h>
@@ -616,4 +616,4 @@ int tt_device_reset(tt_device_t* dev, uint32_t reset_flags);
 }
 #endif
 
-#endif  // TTKMD_H_
+#endif  // UMD_UAPI_H_
